+++ pages/form-builder/preview/settings.tsx
@@ -1,8 +1,8 @@
 import React, { ReactElement } from "react";
 import { useTranslation } from "next-i18next";
 import { PageProps } from "@lib/types";
-import { getServerSideProps } from "./index";
-import { NextPageWithLayout } from "../_app";
+import { getServerSideProps } from "../index";
+import { NextPageWithLayout } from "../../_app";
 import {
   Settings,
   PreviewNavigation,
