+++ pages/form-builder/preview/test-data-delivery.tsx
@@ -1,9 +1,9 @@
 import React, { ReactElement } from "react";
 import { useTranslation } from "next-i18next";
 
-import { NextPageWithLayout } from "../_app";
+import { NextPageWithLayout } from "../../_app";
 import { PageProps } from "@lib/types";
-import { getServerSideProps } from "./index";
+import { getServerSideProps } from "../index";
 import {
   PreviewNavigation,
   TestDataDelivery,
